3d_object_detection_desktop/requirements.txt
fastapi
uvicorn
pymongo
jinja2
aiofiles
python-multipart
python-dotenv
torch
torchvision
numpy
opencv-python
Pillow
deepface
ultralytics